Refactor services and enhance timeline functionality
- Updated `UserExtensions.cs` to check for hyphens in user IDs.
- Initialized properties in `TodoItemsRequest.cs` and set default values.
- Reflected database schema changes in `MediaDbContextModelSnapshot.cs`.
- Modified test cases in `AccessManagementControllerTests.cs` and `CalendarControllerTests.cs` to align with new method signatures.
- Refactored `NotificationsController.cs` to remove dependency on `IAzureNotifications`.
- Added new methods for timeline handling in `IMediaHttpClient` and related services.
- Introduced `TimelineList` and `TimelineListRequest` classes for improved data handling.
- Enhanced `UserAccessService.cs` for better permission checks.
- Updated front-end views and TypeScript files to reflect API changes.
- Added migration `AddCreatedByToPicture` for tracking creator and modification details.

+++ KinaUnaWeb/Services/HttpClients/FamiliesHttpClient.cs
@@ -90,7 +90,7 @@ namespace KinaUnaWeb.Services.HttpClients
             TokenInfo tokenInfo = await _tokenService.GetValidTokenAsync(signedInUserId);
             _httpClient.SetBearerToken(tokenInfo.AccessToken);
 
-            string accessManagementPath = "/api/AccessManagement/FamiliesUserCanAccessList/" + permissionLevel;
+            string accessManagementPath = "/api/AccessManagement/FamiliesUserCanAccessList/" + (int)permissionLevel;
             HttpResponseMessage familiesResponse = await _httpClient.GetAsync(accessManagementPath);
 
             if (!familiesResponse.IsSuccessStatusCode) return new List<Family>();
